Andamento no check de informacoes do stripe

+++ src/Dima.Core/Requests/Stripe/CreateSessionRequest.cs
@@ -5,5 +5,5 @@ public class CreateSessionRequest : Request
     public string OrderNumber { get; set; } = string.Empty;
     public string ProductTitle { get; set; } = string.Empty;
     public string ProductDescription { get; set; } = string.Empty;
-    public decimal OrderTotal { get; set; }
+    public int OrderTotal { get; set; }
 }
